(1248) Split up organizationTypeModel
